Fix for WFMP-145, Introduce an offline CLI command executor
