+++ core/src/main/java/org/wildfly/plugin/core/Utils.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.wildfly.plugin.core;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+/**
+ *
+ * @author jdenise@redhat.com
+ */
+public class Utils {
+
+    public static boolean isValidHomeDirectory(final Path path) {
+        return path != null
+                && Files.exists(path)
+                && Files.isDirectory(path)
+                && Files.exists(path.resolve("jboss-modules.jar"));
+    }
+}
